募集一覧ページとapiを繋ぐ レビュー指摘事項を修正
